% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy.R
\name{copy}
\alias{copy}
\title{Copy Files}
\usage{
copy(from, to, overwrite = FALSE)
}
\arguments{
\item{from}{from path}

\item{to}{to path}

\item{overwrite}{logical, default is FALSE}
}
\description{
Copy Files
}
