% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cut.R
\name{cut_file}
\alias{cut_file}
\title{Cut Files}
\usage{
cut_file(from, to, overwrite = FALSE)
}
\arguments{
\item{from}{from path}

\item{to}{to path}

\item{overwrite}{logical, default is FALSE}
}
\description{
Cut Files
}
