% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lastest.R
\name{lastest_a}
\alias{lastest_a}
\title{Lastest Accessed File}
\usage{
lastest_a(dir = ".", full.names = FALSE, include.dirs = FALSE)
}
\arguments{
\item{dir}{directory}

\item{full.names}{logical, default is FALSE}

\item{include.dirs}{logical, default is FALSE}
}
\value{
file names
}
\description{
Lastest Accessed File
}
\examples{
lastest_a()
}
