% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lastest.R
\name{lastest_c}
\alias{lastest_c}
\title{Lastest Changed File}
\usage{
lastest_c(dir = ".", full.names = FALSE, include.dirs = FALSE)
}
\arguments{
\item{dir}{directory}

\item{full.names}{logical, default is FALSE}

\item{include.dirs}{logical, default is FALSE}
}
\value{
file names
}
\description{
Lastest Changed File
}
\examples{
lastest_c()
}
