% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lastest.R
\name{lastest_m}
\alias{lastest_m}
\title{Lastest Modified File}
\usage{
lastest_m(dir = ".", full.names = FALSE, include.dirs = FALSE)
}
\arguments{
\item{dir}{directory}

\item{full.names}{logical, default is FALSE}

\item{include.dirs}{logical, default is FALSE}
}
\value{
file names
}
\description{
Lastest Modified File
}
\examples{
lastest_m()
}
