% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list.R
\name{list}
\alias{list}
\alias{list_xlsx}
\alias{list_xls}
\alias{list_dir}
\alias{list_docx}
\alias{list_doc}
\alias{list_doc2}
\alias{list_exe}
\alias{list_file}
\alias{list_pdf}
\alias{list_ppt}
\alias{list_pptx}
\alias{list_R}
\alias{list_txt}
\alias{list_tiff}
\alias{list_tif}
\alias{list_png}
\alias{list_jpg}
\alias{list_csv}
\alias{list_zip}
\title{List Excel in Directory}
\usage{
list_xlsx(dir = ".", full.names = FALSE, include.dirs = FALSE)

list_xls(dir = ".", full.names = FALSE, include.dirs = FALSE)

list_dir(dir = ".", full.names = FALSE, include.dirs = FALSE)

list_docx(dir = ".", full.names = FALSE, include.dirs = FALSE)

list_doc(dir = ".", full.names = FALSE, include.dirs = FALSE)

list_doc2(dir = ".", full.names = FALSE, include.dirs = FALSE)

list_exe(dir = ".", full.names = FALSE, include.dirs = FALSE)

list_file(dir = ".", full.names = FALSE, include.dirs = FALSE, pattern = NULL)

list_pdf(dir = ".", full.names = FALSE, include.dirs = FALSE)

list_ppt(dir = ".", full.names = FALSE, include.dirs = FALSE)

list_pptx(dir = ".", full.names = FALSE, include.dirs = FALSE)

list_R(dir = ".", full.names = FALSE, include.dirs = FALSE)

list_txt(dir = ".", full.names = FALSE, include.dirs = FALSE)

list_tiff(dir = ".", full.names = FALSE, include.dirs = FALSE)

list_tif(dir = ".", full.names = FALSE, include.dirs = FALSE)

list_png(dir = ".", full.names = FALSE, include.dirs = FALSE)

list_jpg(dir = ".", full.names = FALSE, include.dirs = FALSE)

list_csv(dir = ".", full.names = FALSE, include.dirs = FALSE)

list_zip(dir = ".", full.names = FALSE, include.dirs = FALSE)
}
\arguments{
\item{dir}{directory}

\item{full.names}{logical, default is FALSE}

\item{include.dirs}{logical, default is FALSE}

\item{pattern}{pattern appendix}
}
\value{
file names
}
\description{
List Excel in Directory
}
\examples{
list_xlsx()
}
