% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list2.R
\name{list2}
\alias{list2}
\alias{list2_xlsx}
\alias{list2_xls}
\alias{list2_dir}
\alias{list2_docx}
\alias{list2_exe}
\alias{list2_file}
\alias{list2_pdf}
\alias{list2_ppt}
\alias{list2_pptx}
\alias{list2_R}
\alias{list2_txt}
\alias{list2_tiff}
\alias{list2_tif}
\alias{list2_png}
\alias{list2_jpg}
\alias{list2_csv}
\alias{list2_zip}
\title{List Excel in Current Directory and Sub-directory}
\usage{
list2_xlsx(dir = ".", full.names = FALSE, include.dirs = FALSE)

list2_xls(dir = ".", full.names = FALSE, include.dirs = FALSE)

list2_dir(dir = ".", full.names = FALSE)

list2_docx(dir = ".", full.names = FALSE, include.dirs = FALSE)

list2_exe(dir = ".", full.names = FALSE, include.dirs = FALSE)

list2_file(dir = ".", full.names = FALSE, include.dirs = FALSE, pattern = NULL)

list2_pdf(dir = ".", full.names = FALSE, include.dirs = FALSE)

list2_ppt(dir = ".", full.names = FALSE, include.dirs = FALSE)

list2_pptx(dir = ".", full.names = FALSE, include.dirs = FALSE)

list2_R(dir = ".", full.names = FALSE, include.dirs = FALSE)

list2_txt(dir = ".", full.names = FALSE, include.dirs = FALSE)

list2_tiff(dir = ".", full.names = FALSE, include.dirs = FALSE)

list2_tif(dir = ".", full.names = FALSE, include.dirs = FALSE)

list2_png(dir = ".", full.names = FALSE, include.dirs = FALSE)

list2_jpg(dir = ".", full.names = FALSE, include.dirs = FALSE)

list2_csv(dir = ".", full.names = FALSE, include.dirs = FALSE)

list2_zip(dir = ".", full.names = FALSE, include.dirs = FALSE)
}
\arguments{
\item{dir}{directory}

\item{full.names}{logical, default is FALSE}

\item{include.dirs}{logical, default is FALSE}

\item{pattern}{pattern appendix}
}
\value{
file names
}
\description{
List Excel in Current Directory and Sub-directory
}
\examples{
list2_xlsx()
}
