% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toUTF8.R
\name{toUTF8}
\alias{toUTF8}
\title{Conert File to UTF8 Coding}
\usage{
toUTF8(file, extension = NULL)
}
\arguments{
\item{file}{files}

\item{extension}{extension fo files. Default is NULL.}
}
\value{
UTF8-coded files.
}
\description{
toUTF8() can be used to convert file coding to UTF8 in windows system.
}
